git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@2602 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ ipstreams/wvresolver.cc
@@ -32,7 +32,7 @@ public:
     ~WvResolverHost()
         {
             if (loop) delete loop;
-            if (pid)
+            if (pid && pid != -1)
             {
                 kill(pid, SIGKILL);
                 waitpid(pid, NULL, 0);
